ZENKO-2731 create location
