+++ src/react/actions/__tests__/utils/testUtil.js
@@ -1,5 +1,5 @@
 // @flow
-import { ErrorMockManagementClient, account, latestOverlay } from '../../../../js/mock/managementClient';
+import { ErrorMockManagementClient, account, latestOverlay, location } from '../../../../js/mock/managementClient';
 import { ErrorUserManager, MockUserManager } from '../../../../js/mock/userManager';
 import { ApiErrorObject } from '../../../../js/mock/error';
 import type { AppState } from '../../../../types/state';
@@ -44,6 +44,7 @@ export const MANAGEMENT_ERROR = new ApiErrorObject(MANAGEMENT_ERROR_MSG, 500);
 
 export const LATEST_OVERLAY = latestOverlay;
 export const ACCOUNT = account;
+export const LOCATION = location;
 
 export function errorUserManagerState(): AppState {
     const state = initState;
