+++ src/react/backend/location/LocationDetails/__tests__/LocationDetailsAws.test.jsx
@@ -1,48 +1,47 @@
-import { checkBox, updateInputText } from '../../../../utils/test';
-import LocationDetailsDOSpaces from '../LocationDetailsDOSpaces';
+import { checkBox, themeMount as mount, updateInputText } from '../../../../utils/test';
+import LocationDetailsAws from '../LocationDetailsAws';
 import React from 'react';
-import { mount } from 'enzyme';
 
 const props = {
     details: {},
     onChange: () => {},
 };
 
-describe('class <LocationDetailsDOSpaces />', () => {
+describe('class <LocationDetailsAws />', () => {
     it('should call onChange on mount', () => {
         const onChangeFn = jest.fn();
         mount(
-            <LocationDetailsDOSpaces {...props} onChange={onChangeFn} />
+            <LocationDetailsAws {...props} onChange={onChangeFn} />
         );
         expect(onChangeFn).toHaveBeenCalledWith({
+            serverSideEncryption: false,
             bucketMatch: false,
             accessKey: '',
             secretKey: '',
             bucketName: '',
-            endpoint: '',
         });
     });
 
     it('should call onChange on state update', () => {
         const refLocation = {
-            endpoint: 'https://ep',
             secretKey: 'sk',
             accessKey: 'ak',
             bucketName: 'bn',
             bucketMatch: true,
+            serverSideEncryption: true,
         };
         const onChangeFn = jest.fn();
-        const component = mount(
-            <LocationDetailsDOSpaces {...props} onChange={onChangeFn} />
+        const { component } = mount(
+            <LocationDetailsAws {...props} onChange={onChangeFn} />
         );
-        component.setState({...refLocation}, () => {
+        component.find(LocationDetailsAws).setState({ ...refLocation }, () => {
             expect(onChangeFn).toHaveBeenCalledWith(refLocation);
         });
     });
 
-    it('should show spaces details for empty details', () => {
-        const component = mount(
-            <LocationDetailsDOSpaces {...props} />
+    it('should show aws details for empty details', () => {
+        const { component } = mount(
+            <LocationDetailsAws {...props} />
         );
         expect(component.find('input[name="accessKey"]')).toHaveLength(1);
         expect(component.find('input[name="accessKey"]').props().value).toEqual('');
@@ -53,23 +52,23 @@ describe('class <LocationDetailsDOSpaces />', () => {
         expect(component.find('input[name="bucketName"]')).toHaveLength(1);
         expect(component.find('input[name="bucketName"]').props().value).toEqual('');
 
-        expect(component.find('input[name="endpoint"]')).toHaveLength(1);
-        expect(component.find('input[name="endpoint"]').props().value).toEqual('');
-
         expect(component.find('input[name="bucketMatch"]')).toHaveLength(1);
         expect(component.find('input[name="bucketMatch"]').props().value).toEqual(false);
+
+        expect(component.find('input[name="serverSideEncryption"]')).toHaveLength(1);
+        expect(component.find('input[name="serverSideEncryption"]').props().value).toEqual(false);
     });
 
-    it('should show spaces details when editing an existing location', () => {
+    it('should show aws details when editing an existing location', () => {
         const locationDetails = {
-            endpoint: 'https://ep',
             secretKey: 'sk',
             accessKey: 'ak',
             bucketName: 'bn',
             bucketMatch: true,
+            serverSideEncryption: true,
         };
-        const component = mount(
-            <LocationDetailsDOSpaces {...props} details={locationDetails} />
+        const { component } = mount(
+            <LocationDetailsAws {...props} details={locationDetails} />
         );
         expect(component.find('input[name="accessKey"]')).toHaveLength(1);
         expect(component.find('input[name="accessKey"]').props().value).toEqual('ak');
@@ -81,30 +80,30 @@ describe('class <LocationDetailsDOSpaces />', () => {
         expect(component.find('input[name="bucketName"]')).toHaveLength(1);
         expect(component.find('input[name="bucketName"]').props().value).toEqual('bn');
 
-        expect(component.find('input[name="endpoint"]')).toHaveLength(1);
-        expect(component.find('input[name="endpoint"]').props().value).toEqual('https://ep');
-
         expect(component.find('input[name="bucketMatch"]')).toHaveLength(1);
         expect(component.find('input[name="bucketMatch"]').props().value).toEqual(true);
+
+        expect(component.find('input[name="serverSideEncryption"]')).toHaveLength(1);
+        expect(component.find('input[name="serverSideEncryption"]').props().value).toEqual(true);
     });
 
     it('should call onChange on location details updates', () => {
         const refLocation = {
-            endpoint: 'https://ep',
             secretKey: 'sk',
             accessKey: 'ak',
             bucketName: 'bn',
             bucketMatch: true,
+            serverSideEncryption: true,
         };
         let location = {};
-        const component = mount(
-            <LocationDetailsDOSpaces {...props} onChange={l => location = l} />
+        const { component } = mount(
+            <LocationDetailsAws {...props} onChange={l => location = l} />
         );
         checkBox(component, 'bucketMatch', true);
+        checkBox(component, 'serverSideEncryption', true);
         updateInputText(component, 'accessKey', 'ak');
         updateInputText(component, 'secretKey', 'sk');
         updateInputText(component, 'bucketName', 'bn');
-        updateInputText(component, 'endpoint', 'https://ep');
 
         expect(location).toEqual(refLocation);
     });
