+++ src/react/backend/location/LocationDetails/__tests__/LocationDetailsAwsCustom.test.jsx
@@ -1,48 +1,56 @@
-import { checkBox, updateInputText } from '../../../../utils/test';
-import LocationDetailsGcp from '../LocationDetailsGcp';
+import { checkBox, themeMount as mount, updateInputText } from '../../../../utils/test';
+import LocationDetailsAwsCustom from '../LocationDetailsAwsCustom';
 import React from 'react';
-import { mount } from 'enzyme';
 
 const props = {
     details: {},
     onChange: () => {},
+    locationType: 'location-scality-ring-s3-v1',
+    capabilities: {
+        locationTypeCephRadosGW: true,
+        locationTypeDigitalOcean: true,
+        locationTypeS3Custom: true,
+        locationTypeSproxyd: true,
+        locationTypeHyperdriveV2: true,
+        locationTypeNFS: true,
+    },
 };
 
-describe('class <LocationDetailsGcp />', () => {
+describe('class <LocationDetailsAwsCustom />', () => {
     it('should call onChange on mount', () => {
         const onChangeFn = jest.fn();
         mount(
-            <LocationDetailsGcp {...props} onChange={onChangeFn} />
+            <LocationDetailsAwsCustom {...props} onChange={onChangeFn} />
         );
         expect(onChangeFn).toHaveBeenCalledWith({
             bucketMatch: false,
             accessKey: '',
             secretKey: '',
             bucketName: '',
-            mpuBucketName: '',
+            endpoint: '',
         });
     });
 
     it('should call onChange on state update', () => {
         const refLocation = {
-            bucketMatch: true,
+            endpoint: 'https://ep',
             secretKey: 'sk',
             accessKey: 'ak',
             bucketName: 'bn',
-            mpuBucketName: 'mbn',
+            bucketMatch: true,
         };
         const onChangeFn = jest.fn();
-        const component = mount(
-            <LocationDetailsGcp {...props} onChange={onChangeFn} />
+        const { component } = mount(
+            <LocationDetailsAwsCustom {...props} onChange={onChangeFn} />
         );
-        component.setState({...refLocation}, () => {
+        component.find(LocationDetailsAwsCustom).setState({ ...refLocation }, () => {
             expect(onChangeFn).toHaveBeenCalledWith(refLocation);
         });
     });
 
-    it('should show gcp details for empty details', () => {
-        const component = mount(
-            <LocationDetailsGcp {...props} />
+    it('should show custom details for empty details', () => {
+        const { component } = mount(
+            <LocationDetailsAwsCustom {...props} />
         );
         expect(component.find('input[name="accessKey"]')).toHaveLength(1);
         expect(component.find('input[name="accessKey"]').props().value).toEqual('');
@@ -53,23 +61,23 @@ describe('class <LocationDetailsGcp />', () => {
         expect(component.find('input[name="bucketName"]')).toHaveLength(1);
         expect(component.find('input[name="bucketName"]').props().value).toEqual('');
 
-        expect(component.find('input[name="mpuBucketName"]')).toHaveLength(1);
-        expect(component.find('input[name="mpuBucketName"]').props().value).toEqual('');
+        expect(component.find('input[name="endpoint"]')).toHaveLength(1);
+        expect(component.find('input[name="endpoint"]').props().value).toEqual('');
 
         expect(component.find('input[name="bucketMatch"]')).toHaveLength(1);
         expect(component.find('input[name="bucketMatch"]').props().value).toEqual(false);
     });
 
-    it('should show gcp details when editing an existing location', () => {
+    it('should show custom details when editing an existing location', () => {
         const locationDetails = {
+            endpoint: 'https://ep',
             secretKey: 'sk',
             accessKey: 'ak',
             bucketName: 'bn',
-            mpuBucketName: 'mbn',
             bucketMatch: true,
         };
-        const component = mount(
-            <LocationDetailsGcp {...props} details={locationDetails} />
+        const { component } = mount(
+            <LocationDetailsAwsCustom {...props} details={locationDetails} />
         );
         expect(component.find('input[name="accessKey"]')).toHaveLength(1);
         expect(component.find('input[name="accessKey"]').props().value).toEqual('ak');
@@ -81,8 +89,8 @@ describe('class <LocationDetailsGcp />', () => {
         expect(component.find('input[name="bucketName"]')).toHaveLength(1);
         expect(component.find('input[name="bucketName"]').props().value).toEqual('bn');
 
-        expect(component.find('input[name="mpuBucketName"]')).toHaveLength(1);
-        expect(component.find('input[name="mpuBucketName"]').props().value).toEqual('mbn');
+        expect(component.find('input[name="endpoint"]')).toHaveLength(1);
+        expect(component.find('input[name="endpoint"]').props().value).toEqual('https://ep');
 
         expect(component.find('input[name="bucketMatch"]')).toHaveLength(1);
         expect(component.find('input[name="bucketMatch"]').props().value).toEqual(true);
@@ -90,21 +98,21 @@ describe('class <LocationDetailsGcp />', () => {
 
     it('should call onChange on location details updates', () => {
         const refLocation = {
+            endpoint: 'https://ep',
             secretKey: 'sk',
             accessKey: 'ak',
             bucketName: 'bn',
-            mpuBucketName: 'mbn',
             bucketMatch: true,
         };
         let location = {};
-        const component = mount(
-            <LocationDetailsGcp {...props} onChange={l => location = l} />
+        const { component } = mount(
+            <LocationDetailsAwsCustom {...props} onChange={l => location = l} />
         );
         checkBox(component, 'bucketMatch', true);
         updateInputText(component, 'accessKey', 'ak');
         updateInputText(component, 'secretKey', 'sk');
         updateInputText(component, 'bucketName', 'bn');
-        updateInputText(component, 'mpuBucketName', 'mbn');
+        updateInputText(component, 'endpoint', 'https://ep');
 
         expect(location).toEqual(refLocation);
     });
