+++ src/react/backend/location/LocationDetails/__tests__/LocationDetailsHyperdriveV2.test.jsx
@@ -27,7 +27,7 @@ describe('class <LocationDetailsHyperdriveV2 />', () => {
         const component = mount(
             <LocationDetailsHyperdriveV2 {...props} onChange={onChangeFn} />
         );
-        component.setState({...refLocation}, () => {
+        component.setState({ ...refLocation }, () => {
             expect(onChangeFn).toHaveBeenCalledWith(refLocation);
         });
     });
