+++ src/react/backend/location/LocationDetails/__tests__/LocationDetailsWasabi.test.jsx
@@ -1,48 +1,47 @@
-import { checkBox, updateInputText } from '../../../../utils/test';
-import LocationDetailsAzure from '../LocationDetailsAzure';
+import { checkBox, themeMount as mount, updateInputText } from '../../../../utils/test';
+import LocationDetailsWasabi from '../LocationDetailsWasabi';
 import React from 'react';
-import { mount } from 'enzyme';
 
 const props = {
     details: {},
     onChange: () => {},
 };
 
-describe('class <LocationDetailsAzure />', () => {
+describe('class <LocationDetailsWasabi />', () => {
     it('should call onChange on mount', () => {
         const onChangeFn = jest.fn();
         mount(
-            <LocationDetailsAzure {...props} onChange={onChangeFn} />
+            <LocationDetailsWasabi {...props} onChange={onChangeFn} />
         );
         expect(onChangeFn).toHaveBeenCalledWith({
             bucketMatch: false,
             accessKey: '',
             secretKey: '',
             bucketName: '',
-            endpoint: '',
+            endpoint: 'https://s3.wasabisys.com',
         });
     });
 
     it('should call onChange on state update', () => {
         const refLocation = {
-            bucketMatch: false,
+            bucketMatch: true,
             accessKey: 'ak',
             secretKey: 'sk',
             bucketName: 'bn',
-            endpoint: 'https://ep',
+            endpoint: 'https://s3.wasabisys.com',
         };
         const onChangeFn = jest.fn();
-        const component = mount(
-            <LocationDetailsAzure {...props} onChange={onChangeFn} />
+        const { component } = mount(
+            <LocationDetailsWasabi {...props} onChange={onChangeFn} />
         );
-        component.setState({...refLocation}, () => {
+        component.find(LocationDetailsWasabi).setState({ ...refLocation }, () => {
             expect(onChangeFn).toHaveBeenCalledWith(refLocation);
         });
     });
 
-    it('should show azure details for empty details', () => {
-        const component = mount(
-            <LocationDetailsAzure {...props} />
+    it('should show wasabi details for empty details', () => {
+        const { component } = mount(
+            <LocationDetailsWasabi {...props} />
         );
         expect(component.find('input[name="accessKey"]')).toHaveLength(1);
         expect(component.find('input[name="accessKey"]').props().value).toEqual('');
@@ -53,29 +52,26 @@ describe('class <LocationDetailsAzure />', () => {
         expect(component.find('input[name="bucketName"]')).toHaveLength(1);
         expect(component.find('input[name="bucketName"]').props().value).toEqual('');
 
+        expect(component.find('input[name="endpoint"]')).toHaveLength(1);
+        expect(component.find('input[name="endpoint"]').props().value).toEqual('https://s3.wasabisys.com');
+
         expect(component.find('input[name="bucketMatch"]')).toHaveLength(1);
         expect(component.find('input[name="bucketMatch"]').props().value).toEqual(false);
-
-        expect(component.find('input[name="endpoint"]')).toHaveLength(1);
-        expect(component.find('input[name="endpoint"]').props().value).toEqual('');
     });
 
-    it('should show azure details when editing an existing location', () => {
+    it('should show custom details when editing an existing location', () => {
         const locationDetails = {
+            endpoint: 'https://s3.wasabisys.com',
             secretKey: 'sk',
             accessKey: 'ak',
             bucketName: 'bn',
             bucketMatch: true,
-            endpoint: 'https://ep',
         };
-        const component = mount(
-            <LocationDetailsAzure {...props} details={locationDetails} />
+        const { component } = mount(
+            <LocationDetailsWasabi {...props} details={locationDetails} />
         );
         expect(component.find('input[name="accessKey"]')).toHaveLength(1);
         expect(component.find('input[name="accessKey"]').props().value).toEqual('ak');
-
-        expect(component.find('input[name="endpoint"]')).toHaveLength(1);
-        expect(component.find('input[name="endpoint"]').props().value).toEqual('https://ep');
 
         expect(component.find('input[name="secretKey"]')).toHaveLength(1);
         // for now we just set it as empty since it's encrypted
@@ -84,21 +80,24 @@ describe('class <LocationDetailsAzure />', () => {
         expect(component.find('input[name="bucketName"]')).toHaveLength(1);
         expect(component.find('input[name="bucketName"]').props().value).toEqual('bn');
 
+        expect(component.find('input[name="endpoint"]')).toHaveLength(1);
+        expect(component.find('input[name="endpoint"]').props().value).toEqual('https://s3.wasabisys.com');
+
         expect(component.find('input[name="bucketMatch"]')).toHaveLength(1);
         expect(component.find('input[name="bucketMatch"]').props().value).toEqual(true);
     });
 
     it('should call onChange on location details updates', () => {
         const refLocation = {
+            endpoint: 'https://s3.wasabisys.com',
             secretKey: 'sk',
             accessKey: 'ak',
             bucketName: 'bn',
             bucketMatch: true,
-            endpoint: 'https://ep',
         };
         let location = {};
-        const component = mount(
-            <LocationDetailsAzure {...props} onChange={l => location = l} />
+        const { component } = mount(
+            <LocationDetailsWasabi {...props} onChange={l => location = l} />
         );
         checkBox(component, 'bucketMatch', true);
         updateInputText(component, 'accessKey', 'ak');
