+++ src/react/backend/location/LocationDetails/index.js
