+++ src/react/backend/location/LocationDetails/storageOptions.js
