+++ src/react/databrowser/BucketCreate.jsx
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import { createBucket } from '../actions';
 import { locationWithIngestion } from '../utils/storageOptions';
 import { push } from 'connected-react-router';
-import { storageOptions } from '../monitor/location/LocationDetails';
+import { storageOptions } from '../backend/location/LocationDetails';
 import styled from 'styled-components';
 
 const SelectOption = styled.div`
